+++ webapp-backend/src/main/java/org/oransc/ric/portal/dashboard/controller/AcXappController.java
@@ -21,6 +21,7 @@ package org.oransc.ric.portal.dashboard.controller;
 
 import java.lang.invoke.MethodHandles;
 
+import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
 import org.oransc.ric.a1med.client.api.A1MediatorApi;
@@ -44,20 +45,19 @@ import io.swagger.annotations.ApiOperation;
 import io.swagger.annotations.ApiParam;
 
 /**
- * * Proxies calls from the front end to the AC xApp via the A1 Mediator API.
- * All methods answer 502 on failure: <blockquote>HTTP server received an
- * invalid response from a server it consulted when acting as a proxy or
+ * Proxies calls from the front end to the AC xApp via the A1 Mediator API. All
+ * methods answer 502 on failure: <blockquote>HTTP server received an invalid
+ * response from a server it consulted when acting as a proxy or
  * gateway.</blockquote>
  */
 @RestController
 @RequestMapping(value = AcXappController.CONTROLLER_PATH, produces = MediaType.APPLICATION_JSON_VALUE)
-public class AcXappController {
+public class AcXappController extends AbstractController {
 
 	private static final Logger logger = LoggerFactory.getLogger(MethodHandles.lookup().lookupClass());
 
 	// Publish paths in constants so tests are easy to write
 	public static final String CONTROLLER_PATH = DashboardConstants.ENDPOINT_PREFIX + "/xapp/ac";
-	// Endpoints
 	public static final String ADMCTRL_METHOD = "/admctrl";
 	public static final String VERSION_METHOD = DashboardConstants.VERSION_METHOD;
 
@@ -78,6 +78,7 @@ public class AcXappController {
 	@ApiOperation(value = "Gets the A1 client library MANIFEST.MF property Implementation-Version.", response = SuccessTransport.class)
 	@GetMapping(VERSION_METHOD)
 	public SuccessTransport getA1MediatorClientVersion() {
+		// No role requirement
 		return new SuccessTransport(200, DashboardApplication.getImplementationVersion(A1MediatorApi.class));
 	}
 
@@ -86,8 +87,9 @@ public class AcXappController {
 	 */
 	@ApiOperation(value = "Gets the admission control policy for AC xApp via the A1 Mediator")
 	@GetMapping(ADMCTRL_METHOD)
-	public Object getAdmissionControlPolicy(HttpServletResponse response) {
-		logger.debug("getAdmissionControlPolicy");
+	public Object getAdmissionControlPolicy(HttpServletRequest request, HttpServletResponse response) {
+		logger.debug("getAdmissionControlPolic}");
+		checkRolesUnpriv(request);
 		response.setStatus(HttpServletResponse.SC_NOT_IMPLEMENTED);
 		return null;
 	}
@@ -98,9 +100,11 @@ public class AcXappController {
 	 */
 	@ApiOperation(value = "Sets the admission control policy for AC xApp via the A1 Mediator")
 	@PutMapping(ADMCTRL_METHOD)
-	public void setAdmissionControlPolicy(@ApiParam(value = "Admission control policy") @RequestBody JsonNode acPolicy, //
+	public void setAdmissionControlPolicy(HttpServletRequest request,
+			@ApiParam(value = "Admission control policy") @RequestBody JsonNode acPolicy, //
 			HttpServletResponse response) {
 		logger.debug("setAdmissionControlPolicy {}", acPolicy);
+		checkRolesPriv(request);
 		a1MediatorApi.a1ControllerPutHandler(AC_CONTROL_NAME, acPolicy);
 		response.setStatus(a1MediatorApi.getApiClient().getStatusCode().value());
 	}