+++ webapp-backend/src/test/java/org/oransc/ric/portal/dashboard/test/controller/AcXappControllerTest.java
@@ -51,7 +51,7 @@ public class AcXappControllerTest extends AbstractControllerTest {
 	@Test
 	public void getTest() throws IOException {
 		// Always returns 501; surprised that no exception is thrown.
-		URI uri = buildUri(null, AcXappController.CONTROLLER_PATH, AcXappController.ADMCTRL_METHOD);
+		URI uri = buildUri(null, AcXappController.CONTROLLER_PATH + "/" + AcXappController.ADMCTRL_METHOD);
 		logger.info("Invoking {}", uri);
 		ResponseEntity<String> response = restTemplate.exchange(uri, HttpMethod.GET, null, String.class);
 		Assertions.assertTrue(response.getStatusCode().is5xxServerError());
@@ -61,7 +61,7 @@ public class AcXappControllerTest extends AbstractControllerTest {
 	public void putTest() throws IOException {
 		ObjectMapper mapper = new ObjectMapper();
 		JsonNode body = mapper.readTree("{ \"policy\" : true }");
-		URI uri = buildUri(null, AcXappController.CONTROLLER_PATH, AcXappController.ADMCTRL_METHOD);
+		URI uri = buildUri(null, AcXappController.CONTROLLER_PATH + "/" + AcXappController.ADMCTRL_METHOD);
 		HttpEntity<JsonNode> entity = new HttpEntity<>(body);
 		logger.info("Invoking {}", uri);
 		ResponseEntity<Void> voidResponse = restTemplate.exchange(uri, HttpMethod.PUT, entity, Void.class);